se agrego una estructura de control en tiendaDescuento

+++ trabajo_practico2/assets-js/notaPromedio.js
@@ -1,6 +1,6 @@
-let cuat1 = parseFloat(prompt ("Ingrese la nota del primer cuatrimestre"));
-let cuat2 = parseFloat(prompt ("Ingrese la nota del segundo cuatrimestre"));
-let cuat3 = parseFloat(prompt ("Ingrese la nota del tercer cuatrimestre"));
+let cuat1 = parseInt(prompt ("Ingrese la nota del primer cuatrimestre"));
+let cuat2 = parseInt(prompt ("Ingrese la nota del segundo cuatrimestre"));
+let cuat3 = parseInt(prompt ("Ingrese la nota del tercer cuatrimestre"));
 let promedio
 promedio = (cuat1+cuat2+cuat3)/3
 
@@ -11,7 +11,7 @@ if (promedio>=0 && promedio<=3){
 }else if (promedio>=6 && promedio<=7) {
     console.log("Nota Buena")
 } else if (promedio>=8 && promedio<=9){
-    console.log("Nota muy Buena")
+    console.log("Nota muy Buena") 
 } else {
     console.log("Nota Sobresaliente " )
 }
